+++ src/Nix/Fresh.hs
@@ -10,6 +10,7 @@
 
 {-# OPTIONS_GHC -Wno-orphans #-}
 
+
 module Nix.Fresh where
 
 import           Control.Applicative
@@ -50,14 +51,16 @@ instance MonadTrans (FreshIdT i) where
 instance MonadBase b m => MonadBase b (FreshIdT i m) where
   liftBase = FreshIdT . liftBase
 
-instance ( MonadVar m
-         , Eq i
-         , Ord i
-         , Show i
-         , Enum i
-         , Typeable i
-         )
-         => MonadThunkId (FreshIdT i m) where
+instance
+  ( MonadVar m
+  , Eq i
+  , Ord i
+  , Show i
+  , Enum i
+  , Typeable i
+  )
+  => MonadThunkId (FreshIdT i m)
+ where
   type ThunkId (FreshIdT i m) = i
   freshId = FreshIdT $ do
     v <- ask
@@ -69,6 +72,7 @@ runFreshIdT i m = runReaderT (unFreshIdT m) i
 -- Orphan instance needed by Infer.hs and Lint.hs
 
 -- Since there's no forking, it's automatically atomic.
+--  NOTE: MonadAtomicRef (ST s) can be upstreamed to `ref-tf`
 instance MonadAtomicRef (ST s) where
   atomicModifyRef r f = do
     v <- readRef r